@@ -20,19 +20,19 @@
     "check": "svelte-check --tsconfig ./tsconfig.json"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
+    "@tailwindcss/vite": "^4.1.7",
     "@tsconfig/svelte": "^5.0.4",
     "autoprefixer": "^10.4.21",
     "postcss": "^8.5.3",
-    "svelte": "^4.2.19",
-    "svelte-check": "^4.1.5",
-    "tailwindcss": "^3.4.17",
+    "svelte": "^5.33.0",
+    "svelte-check": "^4.2.1",
+    "tailwindcss": "^4.1.7",
     "tslib": "^2.8.1",
-    "typescript": "^5.8.2",
-    "vite": "^5.4.14"
+    "typescript": "^5.8.3",
+    "vite": "^6.3.5"
   },
   "engines": {
     "node": ">=18.17.0"
-  },
-  "dependencies": {}
+  }
 }
