@@ -1,6 +1,2 @@
 /** @type {import('tailwindcss').Config} */
-export default {
-  content: ['./index.html', './src/**/*.{svelte,js,ts,jsx,tsx}'],
-  theme: {},
-  plugins: [],
-};
+export default {};
